Removed redundant cast

+++ bomberman/bomberman.cpp
@@ -40,7 +40,7 @@ int main(int argc, char *argv[])
 		game->Render();
 
 		now = SDL_GetPerformanceCounter();
-		deltaTime = (double)((now - last) * 1000 / (double)SDL_GetPerformanceFrequency());
+		deltaTime = (now - last) * 1000 / (float)SDL_GetPerformanceFrequency();
 
 		if (frameDelay > deltaTime)//apply delay
 			SDL_Delay(frameDelay - deltaTime);
